remove old lottery-egld, overwrite lottery-esdt with its code
