--- contracts/examples/lottery-egld/.gitignore
@@ -1,7 +1,0 @@
-# Generated by Cargo
-# will have compiled files and executables
-/target/
-*/target/
-
-# The erdpy output
-output
