--- contracts/examples/lottery-egld/Cargo.toml
@@ -1,29 +1,0 @@
-[package]
-name = "lottery-egld"
-version = "0.0.0"
-authors = [ "Dorin Iancu <dorin.iancu@elrond.com>",]
-edition = "2018"
-publish = false
-
-[lib]
-path = "src/lib.rs"
-
-[features]
-wasm-output-mode = [ "elrond-wasm-node",]
-
-[dependencies.elrond-wasm]
-version = "0.15.1"
-path = "../../../elrond-wasm"
-
-[dependencies.elrond-wasm-derive]
-version = "0.15.1"
-path = "../../../elrond-wasm-derive"
-
-[dependencies.elrond-wasm-node]
-version = "0.15.1"
-path = "../../../elrond-wasm-node"
-optional = true
-
-[dev-dependencies.elrond-wasm-debug]
-version = "0.15.1"
-path = "../../../elrond-wasm-debug"
