--- contracts/examples/lottery-egld/README.md
@@ -1,8 +1,0 @@
-# Interaction
-
-Deploy & interact with contract:
-
-```
-cd contracts/examples/lottery-egld/interaction
-npx ts-node ./playground.ts
-```
