--- contracts/examples/lottery-egld/abi/Cargo.toml
@@ -1,17 +1,0 @@
-[package]
-name = "lottery-egld-abi"
-version = "0.0.0"
-authors = ["Andrei Marinica <andrei.marinica@elrond.com>"]
-edition = "2018"
-publish = false
-
-[dependencies.lottery-egld]
-path = ".."
-
-[dependencies.elrond-wasm]
-version = "0.15.1"
-path = "../../../../elrond-wasm"
-
-[dependencies.elrond-wasm-debug]
-version = "0.15.1"
-path = "../../../../elrond-wasm-debug"
