--- contracts/examples/lottery-egld/abi/src/main.rs
@@ -1,3 +1,0 @@
-fn main() {
-	elrond_wasm_debug::abi_json::print_abi::<lottery_egld::AbiProvider>();
-}
