--- contracts/examples/lottery-egld/elrond.json
@@ -1,3 +1,0 @@
-{
-    "language": "rust"
-}
