--- contracts/examples/lottery-egld/mandos/buy-all-tickets-different-accounts.scen.json
@@ -1,162 +1,0 @@
-{
-    "name": "buy all tickets different accounts",
-    "steps": [
-        {
-            "step": "externalSteps",
-            "path": "start-all-options-bigger-whitelist.scen.json"
-        },
-        {
-            "step": "scCall",
-            "txId": "buy-ticket-acc1",
-            "tx": {
-                "from": "address:acc1",
-                "to": "address:the_lottery_contract",
-                "value": "100",
-                "function": "buy_ticket",
-                "arguments": [
-                    "str:lottery_name"
-                ],
-                "gasLimit": "100,000,000",
-                "gasPrice": "0"
-            },
-            "expect": {
-                "status": "",
-                "gas": "*",
-                "refund": "*"
-            }
-        },
-        {
-            "step": "scCall",
-            "txId": "buy-ticket-acc2",
-            "tx": {
-                "from": "address:acc2",
-                "to": "address:the_lottery_contract",
-                "value": "100",
-                "function": "buy_ticket",
-                "arguments": [
-                    "str:lottery_name"
-                ],
-                "gasLimit": "100,000,000",
-                "gasPrice": "0"
-            },
-            "expect": {
-                "status": "",
-                "gas": "*",
-                "refund": "*"
-            }
-        },
-        {
-            "step": "scCall",
-            "txId": "buy-ticket-acc3",
-            "tx": {
-                "from": "address:acc3",
-                "to": "address:the_lottery_contract",
-                "value": "100",
-                "function": "buy_ticket",
-                "arguments": [
-                    "str:lottery_name"
-                ],
-                "gasLimit": "100,000,000",
-                "gasPrice": "0"
-            },
-            "expect": {
-                "status": "",
-                "gas": "*",
-                "refund": "*"
-            }
-        },
-        {
-            "step": "scCall",
-            "txId": "buy-ticket-acc4",
-            "tx": {
-                "from": "address:acc4",
-                "to": "address:the_lottery_contract",
-                "value": "100",
-                "function": "buy_ticket",
-                "arguments": [
-                    "str:lottery_name"
-                ],
-                "gasLimit": "100,000,000",
-                "gasPrice": "0"
-            },
-            "expect": {
-                "status": "",
-                "gas": "*",
-                "refund": "*"
-            }
-        },
-        {
-            "step": "scCall",
-            "txId": "buy-ticket-acc5",
-            "tx": {
-                "from": "address:acc5",
-                "to": "address:the_lottery_contract",
-                "value": "100",
-                "function": "buy_ticket",
-                "arguments": [
-                    "str:lottery_name"
-                ],
-                "gasLimit": "100,000,000",
-                "gasPrice": "0"
-            },
-            "expect": {
-                "status": "",
-                "gas": "*",
-                "refund": "*"
-            }
-        },
-        {
-            "step": "checkState",
-            "accounts": {
-                "address:my_address": {
-                    "nonce": "2",
-                    "balance": "1,000,000",
-                    "storage": {}
-                },
-                "address:acc1": {
-                    "nonce": "1",
-                    "balance": "999,900",
-                    "storage": {}
-                },
-                "address:acc2": {
-                    "nonce": "1",
-                    "balance": "999,900",
-                    "storage": {}
-                },
-                "address:acc3": {
-                    "nonce": "1",
-                    "balance": "999,900",
-                    "storage": {}
-                },
-                "address:acc4": {
-                    "nonce": "1",
-                    "balance": "999,900",
-                    "storage": {}
-                },
-                "address:acc5": {
-                    "nonce": "1",
-                    "balance": "999,900",
-                    "storage": {}
-                },
-                "address:the_lottery_contract": {
-                    "nonce": "0",
-                    "balance": "500",
-                    "storage": {
-                        "''lotteryInfo|u32:12|''lottery_name": "u32:1|u8:100|u32:0|u64:123,456|u32:1|u32:2|u8:75|u8:25|u32:5|address:acc1|address:acc2|address:acc3|address:acc4|address:acc5|u32:5|u32:2|u16:500",
-                        "''ticketHolder|u32:12|''lottery_name|0x00000000": "address:acc1",
-                        "''ticketHolder|u32:12|''lottery_name|0x00000001": "address:acc2",
-                        "''ticketHolder|u32:12|''lottery_name|0x00000002": "address:acc3",
-                        "''ticketHolder|u32:12|''lottery_name|0x00000003": "address:acc4",
-                        "''ticketHolder|u32:12|''lottery_name|0x00000004": "address:acc5",
-                        "''numberOfEntriesForUser|u32:12|''lottery_name|address:acc1": "1",
-                        "''numberOfEntriesForUser|u32:12|''lottery_name|address:acc2": "1",
-                        "''numberOfEntriesForUser|u32:12|''lottery_name|address:acc3": "1",
-                        "''numberOfEntriesForUser|u32:12|''lottery_name|address:acc4": "1",
-                        "''numberOfEntriesForUser|u32:12|''lottery_name|address:acc5": "1"
-                    },
-                    "code": "file:../output/lottery-egld.wasm"
-                }
-            }
-        }
-    ]
-}
