--- contracts/examples/lottery-egld/mandos/buy-more-tickets-than-allowed.scen.json
@@ -1,60 +1,0 @@
-{
-    "name": "buy more tickets than allowed",
-    "steps": [
-        {
-            "step": "externalSteps",
-            "path": "buy-ticket-all-options.scen.json"
-        },
-        {
-            "step": "scCall",
-            "txId": "buy-more-tickets-than-allowed",
-            "tx": {
-                "from": "address:acc1",
-                "to": "address:the_lottery_contract",
-                "value": "100",
-                "function": "buy_ticket",
-                "arguments": [
-                    "str:lottery_name"
-                ],
-                "gasLimit": "100,000,000",
-                "gasPrice": "0"
-            },
-            "expect": {
-                "status": "4",
-                "message": "str:Ticket limit exceeded for this lottery!",
-                "gas": "*",
-                "refund": "*"
-            }
-        },
-        {
-            "step": "checkState",
-            "accounts": {
-                "address:my_address": {
-                    "nonce": "2",
-                    "balance": "1,000,000",
-                    "storage": {}
-                },
-                "address:acc1": {
-                    "nonce": "2",
-                    "balance": "999,900",
-                    "storage": {}
-                },
-                "address:acc2": {
-                    "nonce": "0",
-                    "balance": "1,000,000",
-                    "storage": {}
-                },
-                "address:the_lottery_contract": {
-                    "nonce": "0",
-                    "balance": "100",
-                    "storage": {
-                        "''lotteryInfo|u32:12|''lottery_name": "u32:1|u8:100|u32:4|u64:123,456|u32:1|u32:2|u8:75|u8:25|u32:3|address:my_address|address:acc1|address:acc2|u32:1|u32:1|u8:100",
-                        "''ticketHolder|u32:12|''lottery_name|0x00000000": "address:acc1",
-                        "''numberOfEntriesForUser|u32:12|''lottery_name|address:acc1": "1"
-                    },
-                    "code": "file:../output/lottery-egld.wasm"
-                }
-            }
-        }
-    ]
-}
