--- contracts/examples/lottery-egld/mandos/buy-ticket-after-deadline.scen.json
@@ -1,66 +1,0 @@
-{
-    "name": "buy ticket after deadline",
-    "steps": [
-        {
-            "step": "externalSteps",
-            "path": "buy-ticket.scen.json"
-        },
-        {
-            "step": "setState",
-            "currentBlockInfo": {
-                "blockTimestamp": "123,567"
-            }
-        },
-        {
-            "step": "scCall",
-            "txId": "buy-ticket-after-deadline",
-            "tx": {
-                "from": "address:acc2",
-                "to": "address:the_lottery_contract",
-                "value": "100",
-                "function": "buy_ticket",
-                "arguments": [
-                    "str:lottery_name"
-                ],
-                "gasLimit": "100,000,000",
-                "gasPrice": "0"
-            },
-            "expect": {
-                "status": "4",
-                "message": "str:Lottery entry period has ended! Awaiting winner announcement.",
-                "gas": "*",
-                "refund": "*"
-            }
-        },
-        {
-            "step": "checkState",
-            "accounts": {
-                "address:my_address": {
-                    "nonce": "2",
-                    "balance": "1,000,000",
-                    "storage": {}
-                },
-                "address:acc1": {
-                    "nonce": "1",
-                    "balance": "999,900",
-                    "storage": {}
-                },
-                "address:acc2": {
-                    "nonce": "1",
-                    "balance": "1,000,000",
-                    "storage": {}
-                },
-                "address:the_lottery_contract": {
-                    "nonce": "0",
-                    "balance": "100",
-                    "storage": {
-                        "''lotteryInfo|u32:12|''lottery_name": "u32:1|u8:100|u32:1|u64:123,456|u32:800|u32:1|u8:100|u32:0|u32:1|u32:1|u8:100",
-                        "''ticketHolder|u32:12|''lottery_name|0x00000000": "address:acc1",
-                        "''numberOfEntriesForUser|u32:12|''lottery_name|address:acc1": "1"
-                    },
-                    "code": "file:../output/lottery-egld.wasm"
-                }
-            }
-        }
-    ]
-}
