--- contracts/examples/lottery-egld/mandos/buy-ticket-after-determined-winner.scen.json
@@ -1,56 +1,0 @@
-{
-    "name": "buy ticket after announced winner",
-    "steps": [
-        {
-            "step": "externalSteps",
-            "path": "determine-winner-different-ticket-holders-winner-acc1.scen.json"
-        },
-        {
-            "step": "scCall",
-            "txId": "buy-ticket-after-announced-winner",
-            "tx": {
-                "from": "address:acc1",
-                "to": "address:the_lottery_contract",
-                "value": "100",
-                "function": "buy_ticket",
-                "arguments": [
-                    "str:lottery_name"
-                ],
-                "gasLimit": "100,000,000",
-                "gasPrice": "0"
-            },
-            "expect": {
-                "status": "4",
-                "message": "str:Lottery is currently inactive.",
-                "gas": "*",
-                "refund": "*"
-            }
-        },
-        {
-            "step": "checkState",
-            "accounts": {
-                "address:my_address": {
-                    "nonce": "3",
-                    "balance": "1,000,000",
-                    "storage": {}
-                },
-                "address:acc1": {
-                    "nonce": "2",
-                    "balance": "1,000,100",
-                    "storage": {}
-                },
-                "address:acc2": {
-                    "nonce": "1",
-                    "balance": "999,900",
-                    "storage": {}
-                },
-                "address:the_lottery_contract": {
-                    "nonce": "0",
-                    "balance": "0",
-                    "storage": {},
-                    "code": "file:../output/lottery-egld.wasm"
-                }
-            }
-        }
-    ]
-}
