--- contracts/examples/lottery-egld/mandos/buy-ticket-second-lottery.scen.json
@@ -1,61 +1,0 @@
-{
-    "name": "buy ticket 2nd lottery",
-    "steps": [
-        {
-            "step": "externalSteps",
-            "path": "start-second-lottery.scen.json"
-        },
-        {
-            "step": "scCall",
-            "txId": "buy-ticket-2nd-lottery",
-            "tx": {
-                "from": "address:acc1",
-                "to": "address:the_lottery_contract",
-                "value": "500",
-                "function": "buy_ticket",
-                "arguments": [
-                    "str:lottery_$$$$"
-                ],
-                "gasLimit": "100,000,000",
-                "gasPrice": "0"
-            },
-            "expect": {
-                "status": "",
-                "gas": "*",
-                "refund": "*"
-            }
-        },
-        {
-            "step": "checkState",
-            "accounts": {
-                "address:my_address": {
-                    "nonce": "2",
-                    "balance": "1,000,000",
-                    "storage": {}
-                },
-                "address:acc1": {
-                    "nonce": "2",
-                    "balance": "999,500",
-                    "storage": {}
-                },
-                "address:acc2": {
-                    "nonce": "0",
-                    "balance": "1,000,000",
-                    "storage": {}
-                },
-                "address:the_lottery_contract": {
-                    "nonce": "0",
-                    "balance": "500",
-                    "storage": {
-                        "''lotteryInfo|u32:12|''lottery_name": "u32:1|u8:100|u32:2|u64:123,456|u32:800|u32:1|u8:100|u32:0|u32:0|u32:0",
-
-                        "''lotteryInfo|u32:12|''lottery_$$$$": "u32:2|u16:500|u32:4|u64:234,567|u32:800|u32:1|u8:100|u32:0|u32:1|u32:2|u16:500",
-                        "''ticketHolder|u32:12|''lottery_$$$$|0x00000000": "address:acc1",
-                        "''numberOfEntriesForUser|u32:12|''lottery_$$$$|address:acc1": "1"
-                    },
-                    "code": "file:../output/lottery-egld.wasm"
-                }
-            }
-        }
-    ]
-}
