--- contracts/examples/lottery-egld/mandos/determine-winner-different-ticket-holders-winner-acc1.scen.json
@@ -1,61 +1,0 @@
-{
-    "name": "acc1 winner",
-    "steps": [
-        {
-            "step": "externalSteps",
-            "path": "buy-ticket-another-account.scen.json"
-        },
-        {
-            "step": "setState",
-            "currentBlockInfo": {
-                "blockTimestamp": "12345678900"
-            }
-        },
-        {
-            "step": "scCall",
-            "txId": "determine-winner-different-ticket-holder-winner-acc1",
-            "tx": {
-                "from": "address:my_address",
-                "to": "address:the_lottery_contract",
-                "value": "0",
-                "function": "determine_winner",
-                "arguments": [
-                    "str:lottery_name"
-                ],
-                "gasLimit": "100,000,000",
-                "gasPrice": "0"
-            },
-            "expect": {
-                "status": "0",
-                "gas": "*",
-                "refund": "*"
-            }
-        },
-        {
-            "step": "checkState",
-            "accounts": {
-                "address:my_address": {
-                    "nonce": "3",
-                    "balance": "1,000,000",
-                    "storage": {}
-                },
-                "address:acc1": {
-                    "nonce": "1",
-                    "balance": "1,000,100",
-                    "storage": {}
-                },
-                "address:acc2": {
-                    "nonce": "1",
-                    "balance": "999,900",
-                    "storage": {}
-                },
-                "address:the_lottery_contract": {
-                    "nonce": "0",
-                    "balance": "0",
-                    "storage": {},
-                    "code": "file:../output/lottery-egld.wasm"
-                }
-            }
-        }
-    ]
-}
