--- contracts/examples/lottery-egld/mandos/determine-winner-same-ticket-holder.scen.json
@@ -1,62 +1,0 @@
-{
-    "name": "same account both tickets",
-    "steps": [
-        {
-            "step": "externalSteps",
-            "path": "buy-ticket-same-account.scen.json"
-        },
-        {
-            "step": "setState",
-            "currentBlockInfo": {
-                "blockNonce": "123,567",
-                "blockTimestamp": "123456789"
-            }
-        },
-        {
-            "step": "scCall",
-            "txId": "determine-winner-same-ticket-holder",
-            "tx": {
-                "from": "address:my_address",
-                "to": "address:the_lottery_contract",
-                "value": "0",
-                "function": "determine_winner",
-                "arguments": [
-                    "str:lottery_name"
-                ],
-                "gasLimit": "100,000,000",
-                "gasPrice": "0"
-            },
-            "expect": {
-                "status": "0",
-                "gas": "*",
-                "refund": "*"
-            }
-        },
-        {
-            "step": "checkState",
-            "accounts": {
-                "address:my_address": {
-                    "nonce": "3",
-                    "balance": "1,000,000",
-                    "storage": {}
-                },
-                "address:acc1": {
-                    "nonce": "2",
-                    "balance": "1,000,000",
-                    "storage": {}
-                },
-                "address:acc2": {
-                    "nonce": "0",
-                    "balance": "1,000,000",
-                    "storage": {}
-                },
-                "address:the_lottery_contract": {
-                    "nonce": "0",
-                    "balance": "0",
-                    "storage": {},
-                    "code": "file:../output/lottery-egld.wasm"
-                }
-            }
-        }
-    ]
-}
