--- contracts/examples/lottery-egld/mandos/determine-winner-split-prize-pool.scen.json
@@ -1,77 +1,0 @@
-{
-    "name": "determine winner with split prize pool acc5 1st, acc1 2nd",
-    "steps": [
-        {
-            "step": "externalSteps",
-            "path": "buy-all-tickets-different-accounts.scen.json"
-        },
-        {
-            "step": "setState",
-            "currentBlockInfo": {
-                "blockTimestamp": "123456789",
-                "blockNonce": "987654321"
-            }
-        },
-        {
-            "step": "scCall",
-            "txId": "determine-winner-with-split-prize-pool",
-            "tx": {
-                "from": "address:acc4",
-                "to": "address:the_lottery_contract",
-                "value": "0",
-                "function": "determine_winner",
-                "arguments": [
-                    "str:lottery_name"
-                ],
-                "gasLimit": "100,000,000",
-                "gasPrice": "0"
-            },
-            "expect": {
-                "status": "0",
-                "gas": "*",
-                "refund": "*"
-            }
-        },
-        {
-            "step": "checkState",
-            "accounts": {
-                "address:my_address": {
-                    "nonce": "2",
-                    "balance": "1,000,000",
-                    "storage": {}
-                },
-                "address:acc1": {
-                    "nonce": "1",
-                    "balance": "1,000,025",
-                    "storage": {}
-                },
-                "address:acc2": {
-                    "nonce": "1",
-                    "balance": "1,000,275",
-                    "storage": {}
-                },
-                "address:acc3": {
-                    "nonce": "1",
-                    "balance": "999,900",
-                    "storage": {}
-                },
-                "address:acc4": {
-                    "nonce": "2",
-                    "balance": "999,900",
-                    "storage": {}
-                },
-                "address:acc5": {
-                    "nonce": "1",
-                    "balance": "999,900",
-                    "storage": {}
-                },
-                "address:the_lottery_contract": {
-                    "nonce": "0",
-                    "balance": "0",
-                    "storage": {},
-                    "code": "file:../output/lottery-egld.wasm"
-                }
-            }
-        }
-    ]
-}
