--- contracts/examples/lottery-egld/mandos/start-after-announced-winner.scen.json
@@ -1,63 +1,0 @@
-{
-    "name": "start lottery for the second time",
-    "steps": [
-        {
-            "step": "externalSteps",
-            "path": "determine-winner-different-ticket-holders-winner-acc1.scen.json"
-        },
-        {
-            "step": "scCall",
-            "txId": "start limited tickets, fixed deadline. Again.",
-            "tx": {
-                "from": "address:my_address",
-                "to": "address:the_lottery_contract",
-                "value": "0",
-                "function": "start",
-                "arguments": [
-                    "str:lottery_name",
-                    "1000",
-                    "0x01|u32:200",
-                    "0x01|u64:12345678905",
-                    "0x",
-                    "0x",
-                    "0x"
-                ],
-                "gasLimit": "100,000,000",
-                "gasPrice": "0"
-            },
-            "expect": {
-                "status": "",
-                "gas": "*",
-                "refund": "*"
-            }
-        },
-        {
-            "step": "checkState",
-            "accounts": {
-                "address:my_address": {
-                    "nonce": "4",
-                    "balance": "1,000,000",
-                    "storage": {}
-                },
-                "address:acc1": {
-                    "nonce": "1",
-                    "balance": "1,000,100",
-                    "storage": {}
-                },
-                "address:acc2": {
-                    "nonce": "1",
-                    "balance": "999,900",
-                    "storage": {}
-                },
-                "address:the_lottery_contract": {
-                    "nonce": "0",
-                    "balance": "0",
-                    "storage": {
-                        "''lotteryInfo|u32:12|''lottery_name": "u32:2|u16:1000|u32:200|u64:12345678905|u32:800|u32:1|u8:100|u32:0|u32:0|u32:0"
-                    },
-                    "code": "file:../output/lottery-egld.wasm"
-                }
-            }
-        }
-    ]
-}
