--- contracts/examples/lottery-egld/mandos/start-second-lottery.scen.json
@@ -1,64 +1,0 @@
-{
-    "name": "start 2nd with limited tickets and fixed deadline",
-    "steps": [
-        {
-            "step": "externalSteps",
-            "path": "start-limited-tickets-and-fixed-deadline.scen.json"
-        },
-        {
-            "step": "scCall",
-            "txId": "start 2nd limited tickets, fixed deadline",
-            "tx": {
-                "from": "address:acc1",
-                "to": "address:the_lottery_contract",
-                "value": "0",
-                "function": "start",
-                "arguments": [
-                    "str:lottery_$$$$",
-                    "500",
-                    "0x01|u32:5",
-                    "0x01|u64:234,567",
-                    "0x",
-                    "0x",
-                    "0x"
-                ],
-                "gasLimit": "100,000,000",
-                "gasPrice": "0"
-            },
-            "expect": {
-                "status": "",
-                "gas": "*",
-                "refund": "*"
-            }
-        },
-        {
-            "step": "checkState",
-            "accounts": {
-                "address:my_address": {
-                    "nonce": "2",
-                    "balance": "1,000,000",
-                    "storage": {}
-                },
-                "address:acc1": {
-                    "nonce": "1",
-                    "balance": "1,000,000",
-                    "storage": {}
-                },
-                "address:acc2": {
-                    "nonce": "0",
-                    "balance": "1,000,000",
-                    "storage": {}
-                },
-                "address:the_lottery_contract": {
-                    "nonce": "0",
-                    "balance": "0",
-                    "storage": {
-                        "''lotteryInfo|u32:12|''lottery_name": "u32:1|u8:100|u32:2|u64:123,456|u32:800|u32:1|u8:100|u32:0|u32:0|u32:0",
-                        "''lotteryInfo|u32:12|''lottery_$$$$": "u32:2|u16:500|u32:5|u64:234,567|u32:800|u32:1|u8:100|u32:0|u32:0|u32:0"
-                    },
-                    "code": "file:../output/lottery-egld.wasm"
-                }
-            }
-        }
-    ]
-}
