--- contracts/examples/lottery-egld/mandos/start-with-no-options.scen.json
@@ -1,63 +1,0 @@
-{
-    "name": "start with no options",
-    "steps": [
-        {
-            "step": "externalSteps",
-            "path": "lottery-init.scen.json"
-        },
-        {
-            "step": "scCall",
-            "txId": "start with no options",
-            "tx": {
-                "from": "address:my_address",
-                "to": "address:the_lottery_contract",
-                "value": "0",
-                "function": "start",
-                "arguments": [
-                    "str:lottery_name",
-                    "100",
-                    "0x",
-                    "0x",
-                    "0x",
-                    "0x",
-                    "0x"
-                ],
-                "gasLimit": "100,000,000",
-                "gasPrice": "0"
-            },
-            "expect": {
-                "status": "",
-                "gas": "*",
-                "refund": "*"
-            }
-        },
-        {
-            "step": "checkState",
-            "accounts": {
-                "address:my_address": {
-                    "nonce": "2",
-                    "balance": "1,000,000",
-                    "storage": {}
-                },
-                "address:acc1": {
-                    "nonce": "0",
-                    "balance": "1,000,000",
-                    "storage": {}
-                },
-                "address:acc2": {
-                    "nonce": "0",
-                    "balance": "1,000,000",
-                    "storage": {}
-                },
-                "address:the_lottery_contract": {
-                    "nonce": "0",
-                    "balance": "0",
-                    "storage": {
-                        "''lotteryInfo|u32:12|''lottery_name": "u32:1|u8:100|u32:800|u64:2592000|u32:800|u32:1|u8:100|u32:0|u32:0|u32:0"
-                    },
-                    "code": "file:../output/lottery-egld.wasm"
-                }
-            }
-        }
-    ]
-}
