--- contracts/examples/lottery-egld/src/lottery_info.rs
@@ -1,16 +1,0 @@
-use elrond_wasm::api::BigUintApi;
-use elrond_wasm::types::{Address, Vec};
-
-elrond_wasm::derive_imports!();
-
-#[derive(NestedEncode, NestedDecode, TopEncode, TopDecode, TypeAbi)]
-pub struct LotteryInfo<BigUint: BigUintApi> {
-	pub ticket_price: BigUint,
-	pub tickets_left: u32,
-	pub deadline: u64,
-	pub max_entries_per_user: u32,
-	pub prize_distribution: Vec<u8>,
-	pub whitelist: Vec<Address>,
-	pub current_ticket_number: usize,
-	pub prize_pool: BigUint,
-}
