--- contracts/examples/lottery-egld/src/status.rs
@@ -1,8 +1,0 @@
-elrond_wasm::derive_imports!();
-
-#[derive(TopEncode, TopDecode, TypeAbi, PartialEq, Clone, Copy)]
-pub enum Status {
-	Inactive,
-	Running,
-	Ended,
-}
