--- contracts/examples/lottery-egld/wasm/Cargo.toml
@@ -1,28 +1,0 @@
-[package]
-name = "lottery-egld-wasm"
-version = "0.0.0"
-edition = "2018"
-publish = false
-
-[lib]
-crate-type = [ "cdylib",]
-
-[workspace]
-members = ["."]
-
-[profile.release]
-codegen-units = 1
-opt-level = "z"
-lto = true
-debug = false
-panic = "abort"
-
-[dependencies.lottery-egld]
-features = [ "wasm-output-mode",]
-default-features = false
-path = ".."
-
-[dependencies.elrond-wasm-output]
-version = "0.15.1"
-path = "../../../../elrond-wasm-output"
-features = [ "wasm-output-mode",]
