--- contracts/examples/lottery-egld/wasm/src/lib.rs
@@ -1,4 +1,0 @@
-#![no_std]
-
-pub use lottery_egld::*;
-pub use elrond_wasm_output::*;
