+++ contracts/examples/lottery-esdt/interaction/abi.json
