+++ contracts/examples/lottery-esdt/interaction/package.json
