+++ contracts/examples/lottery-esdt/interaction/playground.ts
