+++ contracts/examples/lottery-esdt/src/lib.rs
@@ -11,7 +11,7 @@ use lottery_info::LotteryInfo;
 use random::Random;
 use status::Status;
 
-const PERCENTAGE_TOTAL: u16 = 100;
+const PERCENTAGE_TOTAL: u32 = 100;
 const THIRTY_DAYS_IN_SECONDS: u64 = 60 * 60 * 24 * 30;
 const MAX_TICKETS: u32 = 800;
 
@@ -24,7 +24,6 @@ pub trait Lottery {
 	fn start(
 		&self,
 		lottery_name: BoxedBytes,
-		esdt_token_name: TokenIdentifier,
 		ticket_price: Self::BigUint,
 		opt_total_tickets: Option<u32>,
 		opt_deadline: Option<u64>,
@@ -34,7 +33,6 @@ pub trait Lottery {
 	) -> SCResult<()> {
 		self.start_lottery(
 			lottery_name,
-			esdt_token_name,
 			ticket_price,
 			opt_total_tickets,
 			opt_deadline,
@@ -48,7 +46,6 @@ pub trait Lottery {
 	fn create_lottery_pool(
 		&self,
 		lottery_name: BoxedBytes,
-		esdt_token_name: TokenIdentifier,
 		ticket_price: Self::BigUint,
 		opt_total_tickets: Option<u32>,
 		opt_deadline: Option<u64>,
@@ -58,7 +55,6 @@ pub trait Lottery {
 	) -> SCResult<()> {
 		self.start_lottery(
 			lottery_name,
-			esdt_token_name,
 			ticket_price,
 			opt_total_tickets,
 			opt_deadline,
@@ -71,93 +67,76 @@ pub trait Lottery {
 	fn start_lottery(
 		&self,
 		lottery_name: BoxedBytes,
-		esdt_token_name: TokenIdentifier,
 		ticket_price: Self::BigUint,
 		opt_total_tickets: Option<u32>,
 		opt_deadline: Option<u64>,
 		opt_max_entries_per_user: Option<u32>,
-		// prize distribution not supported in the current version
-		// multiple async calls not supported
-		// it will always be set to default 100% to the winner
-		_opt_prize_distribution: Option<Vec<u8>>,
+		opt_prize_distribution: Option<Vec<u8>>,
 		opt_whitelist: Option<Vec<Address>>,
 	) -> SCResult<()> {
 		require!(!lottery_name.is_empty(), "Name can't be empty!");
-		require!(
-			!esdt_token_name.is_egld(),
-			"Esdt token name can't be empty!"
-		);
 
 		let timestamp = self.blockchain().get_block_timestamp();
-
 		let total_tickets = opt_total_tickets.unwrap_or(MAX_TICKETS);
 		let deadline = opt_deadline.unwrap_or_else(|| timestamp + THIRTY_DAYS_IN_SECONDS);
 		let max_entries_per_user = opt_max_entries_per_user.unwrap_or(MAX_TICKETS);
-		let prize_distribution = [PERCENTAGE_TOTAL as u8].to_vec();
+		let prize_distribution =
+			opt_prize_distribution.unwrap_or_else(|| [PERCENTAGE_TOTAL as u8].to_vec());
 		let whitelist = opt_whitelist.unwrap_or_default();
 
 		require!(
 			self.status(&lottery_name) == Status::Inactive,
 			"Lottery is already active!"
 		);
-
 		require!(ticket_price > 0, "Ticket price must be higher than 0!");
-
 		require!(
 			total_tickets > 0,
 			"Must have more than 0 tickets available!"
 		);
-
 		require!(
 			total_tickets <= MAX_TICKETS,
 			"Only 800 or less total tickets per lottery are allowed!"
 		);
-
 		require!(deadline > timestamp, "Deadline can't be in the past!");
-
 		require!(
 			deadline <= timestamp + THIRTY_DAYS_IN_SECONDS,
 			"Deadline can't be later than 30 days from now!"
 		);
-
 		require!(
 			max_entries_per_user > 0,
 			"Must have more than 0 max entries per user!"
 		);
-
 		require!(
 			self.sum_array(&prize_distribution) == PERCENTAGE_TOTAL,
 			"Prize distribution must add up to exactly 100(%)!"
 		);
 
 		let info = LotteryInfo {
-			esdt_token_name,
 			ticket_price,
 			tickets_left: total_tickets,
 			deadline,
 			max_entries_per_user,
 			prize_distribution,
 			whitelist,
-			current_ticket_number: 0u32,
+			current_ticket_number: 0,
 			prize_pool: Self::BigUint::zero(),
 		};
 
-		self.set_lottery_info(&lottery_name, &info);
+		self.lottery_info(&lottery_name).set(&info);
 
 		Ok(())
 	}
 
 	#[endpoint]
-	#[payable("*")]
+	#[payable("EGLD")]
 	fn buy_ticket(
 		&self,
 		lottery_name: BoxedBytes,
 		#[payment] payment: Self::BigUint,
-		#[payment_token] token: TokenIdentifier,
 	) -> SCResult<()> {
 		match self.status(&lottery_name) {
 			Status::Inactive => sc_error!("Lottery is currently inactive."),
-			Status::Running => self.update_after_buy_ticket(&lottery_name, payment, token),
+			Status::Running => self.update_after_buy_ticket(&lottery_name, &payment),
 			Status::Ended => {
 				sc_error!("Lottery entry period has ended! Awaiting winner announcement.")
 			},
@@ -179,13 +158,13 @@ pub trait Lottery {
 
 	#[view]
 	fn status(&self, lottery_name: &BoxedBytes) -> Status {
-		if self.is_empty_lottery_info(lottery_name) {
+		if self.lottery_info(lottery_name).is_empty() {
 			return Status::Inactive;
 		}
 
-		let info = self.get_lottery_info(&lottery_name);
-
-		if self.blockchain().get_block_timestamp() > info.deadline || info.tickets_left == 0 {
+		let info = self.lottery_info(&lottery_name).get();
+		let current_time = self.blockchain().get_block_timestamp();
+		if current_time > info.deadline || info.tickets_left == 0 {
 			return Status::Ended;
 		}
 
@@ -195,161 +174,144 @@ pub trait Lottery {
 	fn update_after_buy_ticket(
 		&self,
 		lottery_name: &BoxedBytes,
-		payment: Self::BigUint,
-		token: TokenIdentifier,
+		payment: &Self::BigUint,
 	) -> SCResult<()> {
-		let mut info = self.get_lottery_info(&lottery_name);
+		let mut info = self.lottery_info(&lottery_name).get();
 		let caller = self.blockchain().get_caller();
 
 		require!(
 			info.whitelist.is_empty() || info.whitelist.contains(&caller),
 			"You are not allowed to participate in this lottery!"
 		);
+		require!(payment == &info.ticket_price, "Wrong ticket fee!");
 
-		require!(token == info.esdt_token_name, "Wrong esdt token!");
-
-		require!(payment == info.ticket_price, "Wrong ticket fee!");
-
-		let mut entries = self.get_number_of_entries_for_user(&lottery_name, &caller);
-
+		let mut entries = self
+			.number_of_entries_for_user(&lottery_name, &caller)
+			.get();
 		require!(
 			entries < info.max_entries_per_user,
 			"Ticket limit exceeded for this lottery!"
 		);
 
-		self.set_ticket_holder(&lottery_name, info.current_ticket_number, &caller);
+		self.ticket_holder(&lottery_name, info.current_ticket_number)
+			.set(&caller);
 		entries += 1;
 		info.current_ticket_number += 1;
 		info.tickets_left -= 1;
-		info.prize_pool += payment;
 
-		self.set_number_of_entries_for_user(lottery_name, &caller, entries);
-		self.set_lottery_info(lottery_name, &info);
+		let ticket_price = info.ticket_price.clone();
+		info.prize_pool += ticket_price;
+
+		self.number_of_entries_for_user(lottery_name, &caller)
+			.set(&entries);
+		self.lottery_info(lottery_name).set(&info);
 
 		Ok(())
 	}
 
 	fn distribute_prizes(&self, lottery_name: &BoxedBytes) {
-		let mut info = self.get_lottery_info(&lottery_name);
+		let mut info = self.lottery_info(&lottery_name).get();
 		let total_tickets = info.current_ticket_number;
 
-		if info.current_ticket_number > 0 {
-			let mut prev_winning_tickets: Vec<u32> = Vec::new();
-
-			let seed = self.blockchain().get_block_random_seed();
-			let mut rand = Random::new(*seed);
-
-			// if there are less tickets that the distributed prize pool,
-			// the 1st place gets the leftover, maybe could split between the remaining
-			// but this is a rare case anyway and it's not worth the overhead
-			let for_loop_end: usize;
-
-			if total_tickets < info.prize_distribution.len() as u32 {
-				for_loop_end = total_tickets as usize;
-			} else {
-				for_loop_end = info.prize_distribution.len();
-			}
-
-			// distribute to the first place last. Laws of probability say that order doesn't matter.
-			// this is done to mitigate the effects of Self::BigUint division leading to "spare" prize money being left out at times
-			// 1st place will get the spare money instead.
-			for i in (0..for_loop_end).rev() {
-				let mut winning_ticket_id: u32;
-
-				loop {
-					winning_ticket_id = rand.next() % total_tickets;
-
-					if !prev_winning_tickets.contains(&winning_ticket_id) {
-						let winner_address =
-							self.get_ticket_holder(&lottery_name, winning_ticket_id);
-						let prize: Self::BigUint;
-
-						if i != 0 {
-							prize = Self::BigUint::from(info.prize_distribution[i] as u32)
-								* info.prize_pool.clone() / Self::BigUint::from(
-								PERCENTAGE_TOTAL as u32,
-							);
-						} else {
-							prize = info.prize_pool.clone();
-						}
-
-						info.prize_pool -= &prize;
-						prev_winning_tickets.push(winning_ticket_id);
-
-						self.set_lottery_info(lottery_name, &info);
-
-						self.send()
-							.direct(&winner_address, &info.esdt_token_name, &prize, &[]);
-
-						break;
-					}
-				}
-			}
+		if total_tickets == 0 {
+			return;
 		}
+
+		// if there are less tickets than the distributed prize pool,
+		// the 1st place gets the leftover, maybe could split between the remaining
+		// but this is a rare case anyway and it's not worth the overhead
+		let total_winning_tickets = if total_tickets < info.prize_distribution.len() {
+			total_tickets as usize
+		} else {
+			info.prize_distribution.len()
+		};
+		let total_prize = info.prize_pool.clone();
+		let winning_tickets = self.get_distinct_random(0, total_tickets, total_winning_tickets);
+		let percentage_total = Self::BigUint::from(PERCENTAGE_TOTAL);
+
+		// distribute to the first place last. Laws of probability say that order doesn't matter.
+		// this is done to mitigate the effects of BigUint division leading to "spare" prize money being left out at times
+		// 1st place will get the spare money instead.
+		for i in (1..total_winning_tickets).rev() {
+			let winning_ticket_id = winning_tickets[i];
+			let winner_address = self.ticket_holder(&lottery_name, winning_ticket_id).get();
+			let prize = &(&Self::BigUint::from(info.prize_distribution[i] as u32) * &total_prize)
+				/ &percentage_total;
+
+			self.send().direct_egld(
+				&winner_address,
+				&prize,
+				b"You won the lottery! Congratulations!",
+			);
+			info.prize_pool -= prize;
+		}
+
+		// send leftover to first place
+		let first_place_winner = self.ticket_holder(&lottery_name, winning_tickets[0]).get();
+		self.send().direct_egld(
+			&first_place_winner,
+			&info.prize_pool,
+			b"You won the lottery, 1st place! Congratulations!",
+		);
 	}
 
 	fn clear_storage(&self, lottery_name: &BoxedBytes) {
-		let info = self.get_lottery_info(lottery_name);
+		let info = self.lottery_info(lottery_name).get();
 
 		for i in 0..info.current_ticket_number {
-			let addr = self.get_ticket_holder(lottery_name, i);
+			let addr = self.ticket_holder(lottery_name, i).get();
 
-			self.clear_ticket_holder(lottery_name, i);
-			self.clear_number_of_entries_for_user(lottery_name, &addr);
+			self.ticket_holder(lottery_name, i).clear();
+			self.number_of_entries_for_user(lottery_name, &addr).clear();
 		}
 
-		self.clear_lottery_info(lottery_name);
+		self.lottery_info(lottery_name).clear();
 	}
 
-	fn sum_array(&self, array: &[u8]) -> u16 {
-		let mut sum = 0u16; // u16 to protect against overflow
+	fn sum_array(&self, array: &[u8]) -> u32 {
+		let mut sum = 0;
 
 		for &item in array {
-			sum += item as u16;
+			sum += item as u32;
 		}
 
 		sum
 	}
 
+	/// does not check if max - min >= amount, that is the caller's job
+	fn get_distinct_random(&self, min: usize, max: usize, amount: usize) -> Vec<usize> {
+		let mut rand_numbers: Vec<usize> = (min..max).collect();
+		let seed = self.blockchain().get_block_random_seed();
+		let mut rand = Random::new(*seed);
+
+		for i in 0..amount {
+			let rand_index = (rand.next() as usize) % amount;
+			rand_numbers.swap(i, rand_index);
+		}
+
+		rand_numbers
+	}
+
 	// storage
 
-	#[storage_set("lotteryInfo")]
-	fn set_lottery_info(
+	#[view(getLotteryInfo)]
+	#[storage_mapper("lotteryInfo")]
+	fn lottery_info(
 		&self,
 		lottery_name: &BoxedBytes,
-		lottery_info: &LotteryInfo<Self::BigUint>,
-	);
+	) -> SingleValueMapper<Self::Storage, LotteryInfo<Self::BigUint>>;
 
-	#[view(lotteryInfo)]
-	#[storage_get("lotteryInfo")]
-	fn get_lottery_info(&self, lottery_name: &BoxedBytes) -> LotteryInfo<Self::BigUint>;
+	#[storage_mapper("ticketHolder")]
+	fn ticket_holder(
+		&self,
+		lottery_name: &BoxedBytes,
+		ticket_id: usize,
+	) -> SingleValueMapper<Self::Storage, Address>;
 
-	#[storage_is_empty("lotteryInfo")]
-	fn is_empty_lottery_info(&self, lottery_name: &BoxedBytes) -> bool;
-
-	#[storage_clear("lotteryInfo")]
-	fn clear_lottery_info(&self, lottery_name: &BoxedBytes);
-
-	#[storage_set("ticketHolder")]
-	fn set_ticket_holder(&self, lottery_name: &BoxedBytes, ticket_id: u32, ticket_holder: &Address);
-
-	#[storage_get("ticketHolder")]
-	fn get_ticket_holder(&self, lottery_name: &BoxedBytes, ticket_id: u32) -> Address;
-
-	#[storage_clear("ticketHolder")]
-	fn clear_ticket_holder(&self, lottery_name: &BoxedBytes, ticket_id: u32);
-
-	#[storage_set("numberOfEntriesForUser")]
-	fn set_number_of_entries_for_user(
+	#[storage_mapper("numberOfEntriesForUser")]
+	fn number_of_entries_for_user(
 		&self,
 		lottery_name: &BoxedBytes,
 		user: &Address,
-		nr_entries: u32,
-	);
-
-	#[storage_get("numberOfEntriesForUser")]
-	fn get_number_of_entries_for_user(&self, lottery_name: &BoxedBytes, user: &Address) -> u32;
-
-	#[storage_clear("numberOfEntriesForUser")]
-	fn clear_number_of_entries_for_user(&self, lottery_name: &BoxedBytes, user: &Address);
+	) -> SingleValueMapper<Self::Storage, u32>;
 }
